assign7_openagent-sdk/pyproject.toml
[project]
name = "assign3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.2",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.1.1",
]
